Started work on a software rasterizer. Tidied up a few things inc sizing buffers and some const correctness

+++ EngineLoop.h
@@ -32,7 +32,7 @@ private:
 	std::unique_ptr<Camera> m_Camera = nullptr;
 
 //Mouse stuff
-	double m_lastMouseX = SCREEN_WIDTH * 0.5, m_lastMouseY = SCREEN_HEIGHT * 0.5;
+	float m_lastMouseX = SCREEN_WIDTH * 0.5, m_lastMouseY = SCREEN_HEIGHT * 0.5;
 	const float m_sensitivity = 0.5f;
 
 	void HandleMouseMovement(float& offsetX, float& offsetY);
